SCRUM-24 Setup initial folder outline. Separated frontend and backend with separate dependencies. Tests use shared config. Modified other files to fit changes. (able to run ./run test, ./run install, ./run <url_file> and it works). Cleaned up class names for metrics.
