+++ README.md
@@ -19,4 +19,4 @@ Our net score calculation is NetScore = ((0.30)*BusFactor + (0.20)*Correctness +
 This command will also return latency scores for each of these metrics.
 
 # ./run test
-This command will run our test suite located in the Metric Tests folder. Our test suite contains tests for each of our metrics as well as our Metric Manager class and cli_parse file.
+This command will run our test suite located in the Metric Tests folder. Our test suite contains tests for each of our metrics as well as our Metric Manager class and CLIParser file.
