+++ backend/.gitignore
@@ -1,0 +1,7 @@
+node_modules/
+typescript testing
+Sample Files for testing
+.env
+coverage/
+*.log
+*.pyc
