+++ frontend/next.config.mjs
