+++ run
@@ -7,8 +7,16 @@ import json
 
 def install():
     try:
-        subprocess.run(["npm", "install"], check=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-        print(f"Dependencies installed successfully")
+        # Install dependencies for frontend
+        subprocess.run(["npm", "install"], cwd=os.path.join(os.path.dirname(__file__), "frontend"), check=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+
+        # Install dependencies for backend
+        subprocess.run(["npm", "install"], cwd=os.path.join(os.path.dirname(__file__), "backend"), check=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+
+        # Install dependencies for jest
+        subprocess.run(["npm", "install"], cwd=os.path.join(os.path.dirname(__file__)), check=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+        print("Dependencies installed successfully")
+        
     except subprocess.CalledProcessError as e:
         print(f"Error installing dependencies: {e}")
         sys.exit(1)
@@ -20,7 +28,8 @@ def tests():
             ["npx", "jest", "--json", "--coverage", "--silent"],
             text=True,
             capture_output=True,
-            check=True
+            check=True,
+            cwd=os.path.dirname(__file__)
         )
 
         # Deserialize the Jest output from JSON format
@@ -30,11 +39,15 @@ def tests():
 
         # Calculate the coverage percentage
         coverage_data = test_results.get("coverageMap", {})
-        lines_total = sum(len(file_data['statementMap']) for file_data in coverage_data.values())
-        lines_covered = sum(1 for file_data in coverage_data.values() 
-                    for line_key in file_data['statementMap'] 
-                    if file_data['s'].get(line_key, 0) > 0)
-        coverage_percentage = (lines_covered / lines_total * 100) if lines_total > 0 else 0
+
+        if coverage_data:
+            lines_total = sum(len(file_data['statementMap']) for file_data in coverage_data.values())
+            lines_covered = sum(1 for file_data in coverage_data.values() 
+                        for line_key in file_data['statementMap'] 
+                        if file_data['s'].get(line_key, 0) > 0)
+            coverage_percentage = (lines_covered / lines_total * 100) if lines_total > 0 else 0
+        else:
+            coverage_percentage = 0
 
         # Final Test Results
         print(f"Total: {total_tests}")
@@ -59,7 +72,7 @@ def URL_FILE(file_path):
         for url in urls:
             url = url.strip()
             if url:
-                result = subprocess.run(["npx", "ts-node", "./webpage-461/src/cli_parse.ts", url], capture_output=True, text=True)
+                result = subprocess.run(["npx", "ts-node", "./backend/src/services/metrics/CLIParser.ts", url], capture_output=True, text=True, cwd=os.path.dirname(__file__))
                 print(result.stdout)
                 if result.stderr:
                     print(result.stderr, file=sys.stderr)
